requirements/requirements_base.txt
# Core dependencies with exact version pinning for stability
# Last updated: 2024-09-02
# Python version: >=3.11,<3.13

# AI/ML Core Dependencies
pinecone-client[grpc]==5.0.1
langchain==0.2.14
langchain-openai==0.1.22  
langchain-community==0.2.12
langchain-pinecone==0.1.3
langchain-core==0.2.32
langchainhub==0.1.21

# OpenAI Dependencies (pinned for consistency)
openai==1.40.0
tiktoken==0.7.0

# Web Framework Stack
fastapi==0.112.1
uvicorn[standard]==0.30.6
starlette==0.38.2
websockets==13.0
jinja2==3.1.4

# Configuration & Settings
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1

# HTTP & Async
httpx==0.27.0
anyio==4.4.0
sniffio==1.3.1

# Core Python Utilities
typing-extensions==4.12.2
annotated-types==0.7.0

# JSON & Data Processing
orjson==3.10.7
python-multipart==0.0.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging & Monitoring (optional but recommended)
structlog==24.2.0
rich==13.7.1